requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
alabaster=0.7.11=py37_0
anaconda=5.3.0=py37_0
anaconda-client=1.7.2=py37_0
anaconda-project=0.8.2=py37_0
appdirs=1.4.3=py37h28b3542_0
asn1crypto=0.24.0=py37_0
astroid=2.0.4=py37_0
astropy=3.0.4=py37h14c3975_0
atomicwrites=1.2.1=py37_0
attrs=18.2.0=py37h28b3542_0
automat=0.7.0=py37_0
babel=2.6.0=py37_0
backcall=0.1.0=py37_0
backports=1.0=py37_1
backports.shutil_get_terminal_size=1.0.0=py37_2
beautifulsoup4=4.6.3=py37_0
bitarray=0.8.3=py37h14c3975_0
bkcharts=0.2=py37_0
blas=1.0=mkl
blaze=0.11.3=py37_0
bleach=2.1.4=py37_0
blosc=1.14.4=hdbcaa40_0
bokeh=0.13.0=py37_0
boto=2.49.0=py37_0
bottleneck=1.2.1=py37h035aef0_1
bzip2=1.0.6=h14c3975_5
ca-certificates=2018.03.07=0
cairo=1.14.12=h8948797_3
certifi=2018.8.24=py37_1
cffi=1.11.5=py37he75722e_1
chardet=3.0.4=py37_1
click=6.7=py37_0
cloudpickle=0.5.5=py37_0
clyent=1.2.2=py37_1
colorama=0.3.9=py37_0
constantly=15.1.0=py37h28b3542_0
contextlib2=0.5.5=py37_0
cryptography=2.3.1=py37hc365091_0
curl=7.61.0=h84994c4_0
cycler=0.10.0=py37_0
cython=0.28.5=py37hf484d3e_0
cytoolz=0.9.0.1=py37h14c3975_1
dask=0.19.1=py37_0
dask-core=0.19.1=py37_0
datashape=0.5.4=py37_1
dbus=1.13.2=h714fa37_1
decorator=4.3.0=py37_0
defusedxml=0.5.0=py37_1
distributed=1.23.1=py37_0
docutils=0.14=py37_0
entrypoints=0.2.3=py37_2
et_xmlfile=1.0.1=py37_0
expat=2.2.6=he6710b0_0
fastcache=1.0.2=py37h14c3975_2
filelock=3.0.8=py37_0
flask=1.0.2=py37_1
flask-cors=3.0.6=py37_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
fribidi=1.0.5=h7b6447c_0
get_terminal_size=1.0.0=haa9412d_0
gevent=1.3.6=py37h7b6447c_0
glib=2.56.2=hd408876_0
glob2=0.6=py37_0
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py37h10f8cd9_2
graphite2=1.3.12=h23475e2_2
greenlet=0.4.15=py37h7b6447c_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
heapdict=1.0.0=py37_2
html5lib=1.0.1=py37_0
hyperlink=18.0.0=py37_0
icu=58.2=h9c2bf20_1
idna=2.7=py37_0
imageio=2.4.1=py37_0
imagesize=1.1.0=py37_0
incremental=17.5.0=py37_0
intel-openmp=2019.0=118
ipykernel=4.9.0=py37_1
ipython=6.5.0=py37_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.4.1=py37_0
isort=4.3.4=py37_0
itsdangerous=0.24=py37_1
jbig=2.1=hdba287a_0
jdcal=1.4=py37_0
jedi=0.12.1=py37_0
jeepney=0.3.1=py37_0
jinja2=2.10=py37_0
jpeg=9b=h024ee3a_2
jsonschema=2.6.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.2.3=py37_0
jupyter_console=5.2.0=py37_1
jupyter_core=4.4.0=py37_0
jupyterlab=0.34.9=py37_0
jupyterlab_launcher=0.13.1=py37_0
keyring=13.2.1=py37_0
kiwisolver=1.0.1=py37hf484d3e_0
lazy-object-proxy=1.3.1=py37h14c3975_2
libcurl=7.61.0=h1ad7b7a_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.34=hb9fc6fc_0
libsodium=1.0.16=h1bed415_0
libssh2=1.8.0=h9cfc8f7_4
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libtool=2.4.6=h544aabb_3
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
libxslt=1.1.32=h1312cb7_0
llvmlite=0.24.0=py37hdbcaa40_0
locket=0.2.0=py37_1
lxml=4.2.5=py37hefd8a0e_0
lzo=2.10=h49e0be7_2
markupsafe=1.0=py37h14c3975_1
matplotlib=2.2.3=py37hb69df0a_0
mccabe=0.6.1=py37_1
mistune=0.8.3=py37h14c3975_1
mkl=2019.0=118
mkl-service=1.1.2=py37h90e4bf4_5
mkl_fft=1.0.4=py37h4414c95_1
mkl_random=1.0.1=py37h4414c95_1
more-itertools=4.3.0=py37_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpmath=1.0.0=py37_2
msgpack-python=0.5.6=py37h6bb024c_1
multipledispatch=0.6.0=py37_0
nbconvert=5.4.0=py37_1
nbformat=4.4.0=py37_0
ncurses=6.1=hf484d3e_0
networkx=2.1=py37_0
ninja=1.8.2=py37h6bb024c_1
nltk=3.3.0=py37_0
nose=1.3.7=py37_2
notebook=5.6.0=py37_0
numba=0.39.0=py37h04863e7_0
numexpr=2.6.8=py37hd89afb7_0
numpy=1.15.1=py37h1d66e8a_0
numpy-base=1.15.1=py37h81de0dd_0
numpydoc=0.8.0=py37_0
odo=0.5.1=py37_0
olefile=0.46=py37_0
openpyxl=2.5.6=py37_0
openssl=1.0.2p=h14c3975_0
packaging=17.1=py37_0
pandas=0.23.4=py37h04863e7_0
pandoc=1.19.2.1=hea2e7c5_1
pandocfilters=1.4.2=py37_1
pango=1.42.4=h049681c_0
parso=0.3.1=py37_0
partd=0.3.8=py37_0
patchelf=0.9=hf484d3e_2
path.py=11.1.0=py37_0
pathlib2=2.3.2=py37_0
patsy=0.5.0=py37_0
pcre=8.42=h439df22_0
pep8=1.7.1=py37_0
pexpect=4.6.0=py37_0
pickleshare=0.7.4=py37_0
pillow=5.2.0=py37heded4f4_0
pip=10.0.1=py37_0
pixman=0.34.0=hceecf20_3
pkginfo=1.4.2=py37_1
pluggy=0.7.1=py37h28b3542_0
ply=3.11=py37_0
prometheus_client=0.3.1=py37h28b3542_0
prompt_toolkit=1.0.15=py37_0
psutil=5.4.7=py37h14c3975_0
ptyprocess=0.6.0=py37_0
py=1.6.0=py37_0
pyasn1=0.4.4=py37h28b3542_0
pyasn1-modules=0.2.2=py37_0
pycodestyle=2.4.0=py37_0
pycosat=0.6.3=py37h14c3975_0
pycparser=2.18=py37_1
pycrypto=2.6.1=py37h14c3975_9
pycurl=7.43.0.2=py37hb7f436b_0
pyflakes=2.0.0=py37_0
pygments=2.2.0=py37_0
pylint=2.1.1=py37_0
pyodbc=4.0.24=py37he6710b0_0
pyopenssl=18.0.0=py37_0
pyparsing=2.2.0=py37_1
pyqt=5.9.2=py37h05f1152_2
pysocks=1.6.8=py37_0
pytables=3.4.4=py37ha205bf6_0
pytest=3.8.0=py37_0
pytest-arraydiff=0.2=py37h39e3cac_0
pytest-astropy=0.4.0=py37_0
pytest-doctestplus=0.1.3=py37_0
pytest-openfiles=0.3.0=py37_0
pytest-remotedata=0.3.0=py37_0
python=3.7.0=hc3d631a_0
python-dateutil=2.7.3=py37_0
pytorch=0.4.1=py37_py36_py35_py27__9.0.176_7.1.2_2
pytz=2018.5=py37_0
pywavelets=1.0.0=py37hdd07704_0
pyyaml=3.13=py37h14c3975_0
pyzmq=17.1.2=py37h14c3975_0
qt=5.9.6=h8703b6f_2
qtawesome=0.4.4=py37_0
qtconsole=4.4.1=py37_0
qtpy=1.5.0=py37_0
readline=7.0=h7b6447c_5
requests=2.19.1=py37_0
rope=0.11.0=py37_0
ruamel_yaml=0.15.46=py37h14c3975_0
scikit-image=0.14.0=py37hf484d3e_1
scikit-learn=0.19.2=py37h4989274_0
scipy=1.1.0=py37hfa4b5c9_1
seaborn=0.9.0=py37_0
secretstorage=3.1.0=py37_0
send2trash=1.5.0=py37_0
service_identity=17.0.0=py37h28b3542_0
setuptools=40.2.0=py37_0
simplegeneric=0.8.1=py37_2
singledispatch=3.4.0.3=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.11.0=py37_1
snappy=1.1.7=hbae5bb6_3
snowballstemmer=1.2.1=py37_0
sortedcollections=1.0.1=py37_0
sortedcontainers=2.0.5=py37_0
sphinx=1.7.9=py37_0
sphinxcontrib=1.0=py37_1
sphinxcontrib-websupport=1.1.0=py37_1
spyder=3.3.1=py37_1
spyder-kernels=0.2.6=py37_0
sqlalchemy=1.2.11=py37h7b6447c_0
sqlite=3.24.0=h84994c4_0
statsmodels=0.9.0=py37h035aef0_0
sympy=1.2=py37_0
tblib=1.3.2=py37_0
terminado=0.8.1=py37_1
testpath=0.3.1=py37_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=py37_0
torchvision=0.2.1=py37_1
tornado=5.1=py37h14c3975_0
tqdm=4.26.0=py37h28b3542_0
traitlets=4.3.2=py37_0
twisted=18.7.0=py37h14c3975_1
unicodecsv=0.14.1=py37_0
unixodbc=2.3.7=h14c3975_0
urllib3=1.23=py37_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=0.14.1=py37_0
wheel=0.31.1=py37_0
widgetsnbextension=3.4.1=py37_0
wrapt=1.10.11=py37h14c3975_2
xlrd=1.1.0=py37_1
xlsxwriter=1.1.0=py37_0
xlwt=1.3.0=py37_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.2.5=hf484d3e_1
zict=0.1.3=py37_0
zlib=1.2.11=ha838bed_2
zope=1.0=py37_1
zope.interface=4.5.0=py37h14c3975_0
